Moving IT SDI to addon
