+++ addons/addons.go
@@ -11,6 +11,7 @@ import (
 	// Import all the addons to ensure they're ready to use.
 	_ "github.com/invopop/gobl/addons/es/facturae"
 	_ "github.com/invopop/gobl/addons/es/tbai"
+	_ "github.com/invopop/gobl/addons/it/sdi"
 	_ "github.com/invopop/gobl/addons/mx/cfdi"
 	_ "github.com/invopop/gobl/addons/pt/saft"
 )
