+++ data/schemas/tax/addon-def.json
@@ -44,6 +44,14 @@
           "title": "Scenarios",
           "description": "Scenarios are applied to documents after normalization and before\nvalidation to ensure that form specific extensions have been added\nto the document."
         },
+        "inboxes": {
+          "items": {
+            "$ref": "https://gobl.org/draft-0/cbc/key-definition"
+          },
+          "type": "array",
+          "title": "Inboxes",
+          "description": "Inboxes is a list of keys that are used to identify where copies of\ndocuments can be sent."
+        },
         "corrections": {
           "$ref": "#/$defs/CorrectionSet",
           "title": "Corrections",
