+++ regimes/it/categories.go
@@ -19,6 +19,12 @@ const (
 	TaxCategoryENPAM    cbc.Code = "ENPAM"
 )
 
+// Keys used for meta data from external sources.
+const (
+	KeyFatturaPATipoRitenuta      cbc.Key = "fatturapa-tipo-ritenuta"
+	KeyFatturaPAModalitaPagamento cbc.Key = "fatturapa-modalita-pagamento"
+)
+
 var categories = []*tax.CategoryDef{
 	{
 		Code:     tax.CategoryVAT,
@@ -30,9 +36,6 @@ var categories = []*tax.CategoryDef{
 		Title: i18n.String{
 			i18n.EN: "Value Added Tax",
 			i18n.IT: "Imposta sul Valore Aggiunto",
-		},
-		Extensions: []cbc.Key{
-			ExtKeySDIExempt,
 		},
 		Rates: []*tax.RateDef{
 			{
@@ -121,7 +124,6 @@ var categories = []*tax.CategoryDef{
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT01",
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetained},
 	},
 	{
 		Code:     TaxCategoryIRES,
@@ -137,7 +139,6 @@ var categories = []*tax.CategoryDef{
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT02",
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetained},
 	},
 	{
 		Code:     TaxCategoryINPS,
@@ -150,7 +151,6 @@ var categories = []*tax.CategoryDef{
 			i18n.EN: "Contribution to the National Social Security Institute",
 			i18n.IT: "Contributo Istituto Nazionale della Previdenza Sociale", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT03",
 		},
@@ -166,7 +166,6 @@ var categories = []*tax.CategoryDef{
 			i18n.EN: "Contribution to the National Welfare Board for Sales Agents and Representatives",
 			i18n.IT: "Contributo Ente Nazionale Assistenza Agenti e Rappresentanti di Commercio", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT04",
 		},
@@ -182,7 +181,6 @@ var categories = []*tax.CategoryDef{
 			i18n.EN: "Contribution to the National Pension and Welfare Board for Doctors",
 			i18n.IT: "Contributo - Ente Nazionale Previdenza e Assistenza Medici", // nolint:misspell
 		},
-		Extensions: []cbc.Key{ExtKeySDIRetained},
 		Map: cbc.CodeMap{
 			KeyFatturaPATipoRitenuta: "RT05",
 		},
