+++ regimes/it/examples/hotel.yaml
@@ -1,4 +1,5 @@
 $schema: "https://gobl.org/draft-0/bill/invoice"
+$addons: ["it-sdi-v1"]
 uuid: 3aea7b56-59d8-4beb-90bd-f8f280d852a0
 series: SAMPLE
 code: "002"
